reogranize code based on suggestion
